Prevent keyboard rendering over prompt
When a text input field has `autofocus` set, the input can grab focus and cause the keyboard to open. The keyboard then obscures the install prompt and the user must close the keyboard to see the entire prompt.

Since the prompt is meant to be a modal, this commit automatically blurs any element that has focus when it gets rendered, which should fix the problem.

Unfortunately, in my testing, the problem didn’t consistently appear on iOS. But this fix does appear to ensure that the text field that normally would have focus on a page load doesn’t.